Add details on the 2 webhook contexts (#444)
Explain that the GITHUB_WEBHOOK_VALIDATION_CONTEXT and
GITHUB_WEBHOOK_SYNC_CONTEXT can be set to any string and will be used
to as the name of the checks reported to GitHub Actions.

Would be better to have this detail in a more comprehensive doc as
proposed by #434 but this should give those following TESTING.md
some clarity.

Closes: #422

Signed-off-by: Thanh Ha <zxiiro@gmail.com>