Minor optimization in list advisories operation
Previously, a new GitHubProvider was created for each state, when
iterating over states, and fetching security advisories for each state,
which is not necessary.

This commit swaps state for-loop and provider with-statement, to re-use
the same provider for all iterations.

This also allows using the provider outside of the for loop (see
subsequent commits).

Signed-off-by: Lukas Puehringer <lukas.puehringer@eclipse-foundation.org>