Add 'last_commented_at' column to GHSA table
Add new column to the GHSA table in 'list-advisories', which show the
date of the most recent comment on the GHSA.

Unfortunately, this information is not available via REST or GraphQL
APIs, so we make use of otterdog's WebClient.

Using the WebClient requires 4x as many secrets from password store,
which makes the command significantly slower.

Signed-off-by: Lukas Puehringer <lukas.puehringer@eclipse-foundation.org>